requirements.txt
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cymem==2.0.4
doc2text==0.2.4
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
Flask==1.1.2
future==0.18.2
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
logging==0.4.9.6
lxml==4.6.1
MarkupSafe==1.1.1
mime==0.1.0
murmurhash==1.0.4
numpy==1.18.4
Pillow==8.0.1
plac==1.1.3
preshed==3.0.3
PyPDF2==1.26.0
pytesseract==0.3.6
python-docx==0.8.10
PyYAML==5.3.1
requests==2.24.0
scipy==1.5.4
selenium==3.141.0
spacy==2.3.2
srsly==1.0.3
thinc==7.4.1
tqdm==4.51.0
urllib3==1.25.11
wasabi==0.8.0
Werkzeug==1.0.1
zipp==3.4.0
